Removed autogenerate command, because now autogeneration is based on config and doesn't require a command other than default generate. Renamed autogenerate_tsv module to maketsv.
